oh my god why is the volume control linear who even did this the monstronsity

+++ Source/Core/AudioCommon/CoreAudioSoundStream.h
@@ -15,12 +15,13 @@ class CoreAudioSound final : public SoundStream
 #ifdef __APPLE__
 public:
   bool Start() override;
-  void SetVolume(int volume) override;
+  void SetVolume(float volume) override;
   void SoundLoop() override;
   void Stop() override;
   void Update() override;
 
   static bool isValid() { return true; }
+
 private:
   AudioUnit audioUnit;
   int m_volume;
